Add functionality to create JSON files if they dont exist

+++ lib/preserve_data/load_books.rb
@@ -2,9 +2,7 @@ DATA_PATH = './data'.freeze
 
 def load_books
   file_path = "#{DATA_PATH}/books.json"
-  unless File.exist?(file_path)
-    File.write(file_path, '[]')
-  end
+  File.write(file_path, '[]') unless File.exist?(file_path)
 
   json_data = File.read(file_path)
   return if json_data.empty?
